pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "astros"
version = "0.1.0"
description = "AI-Integrated Operating System"
authors = [{name = "AstrOS Team", email = "team@astros.org"}]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "click>=8.1.0",
    "PyYAML>=6.0",
    # AI Dependencies
    "transformers>=4.35.0",
    "torch>=2.1.0",
    "spacy>=3.7.0",
    "nltk>=3.8.0",
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",
    # OpenAI Integration
    "openai>=1.0.0",
    "tiktoken>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.10.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
]

[project.scripts]
astros = "astros.cli:main"
astros-agent = "astros.agent:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88